+++ app/src/services/status-queue.service.js
@@ -1,7 +1,10 @@
 const logger = require('logger');
 const config = require('config');
+const sleep = require('sleep');
 const amqp = require('amqplib');
 const docImporter = require('rw-doc-importer-messages');
+
+let retries = 10;
 
 class StatusQueueService {
 
@@ -11,11 +14,26 @@ class StatusQueueService {
             this.init().then(() => {
                 logger.info('Connected');
             }, (err) => {
-                logger.error(err);
-                process.exit(1);
+                this.retryConnection(err);
             });
         } catch (err) {
             logger.error(err);
+        }
+    }
+
+    retryConnection(err) {
+        if (retries >= 0) {
+            retries--;
+            logger.error(`Failed to connect to RabbitMQ uri ${config.get('rabbitmq.url')} with error message "${err.message}", retrying...`);
+            sleep.sleep(2);
+            this.init().then(() => {
+                logger.info('Connected');
+            }, (initError) => {
+                this.retryConnection(initError);
+            });
+        } else {
+            logger.error(err);
+            process.exit(1);
         }
     }