+++ app/test/e2e/test-server.js
@@ -1,3 +1,4 @@
+const nock = require('nock');
 const chai = require('chai');
 const chaiHttp = require('chai-http');
 
@@ -9,6 +10,11 @@ exports.getTestServer = function getTestServer() {
     if (requester) {
         return requester;
     }
+
+    nock(`http://${process.env.ELASTIC_URL}`)
+        .persist()
+        .head('/')
+        .reply(200);
 
     const server = require('../../src/app');
     requester = chai.request(server).keepOpen();